requirements.txt
ipykernel
transformers
datasets
torch 
scikit-learn
torchvision
matplotlib
#this is the library. for downloading youtube videos
yt-dlp
# for lip detection
mediapipe
opencv-python

# for swinTransformerfromscratch

einops

# for audio trimming
openai-whisper 
stable-ts
pydub 
openpyxl
whisperx
#install sudo ffmpeg